requirements.txt
google-genai
python-dotenv
pytesseract
pillow